[benc/input-number-v2-take-3] Move IN->NI conversion functions into parser
