+++ packages/perseus-core/src/parse-perseus-json/perseus-parsers/input-number-widget.ts
@@ -1,4 +1,3 @@
-import {convertInputNumberOptionsToNumericInput} from "../../widgets/input-number/to-numeric-input";
 import {
     boolean,
     constant,
@@ -18,6 +17,10 @@ import {versionedWidgetOptions} from "./versioned-widget-options";
 import {parseWidgetWithVersion} from "./widget";
 
 import type {ParsedValue, Parser} from "../parser-types";
+import type {
+    MathFormat,
+    PerseusInputNumberAnswerType,
+} from "@khanacademy/perseus-core";
 
 const booleanToZero: Parser<number> = (rawValue, ctx) => {
     if (typeof rawValue === "boolean") {
@@ -42,6 +45,8 @@ export const parseInputNumberWidgetV1 = parseWidgetWithVersion(
     object({
         size: enumeration("normal", "small"),
         coefficient: constant(false),
+        labelText: optional(string),
+        rightAlign: optional(boolean),
         answers: singleton(
             object({
                 value: nullable(number),
@@ -67,8 +72,10 @@ function migrateV0ToV1(
     };
 }
 
+export type PerseusInputNumberWidgetOptionsV0 = ParsedValue<typeof parseInputNumberWidgetV0>["options"];
+
 export const parseInputNumberWidgetV0 = parseWidgetWithVersion(
-    object({major: constant(0), minor: number}),
+    optional(object({major: constant(0), minor: number})),
     constant("input-number"),
     object({
         answerType: optional(
@@ -98,6 +105,78 @@ export const parseInputNumberWidgetV0 = parseWidgetWithVersion(
         ),
     }),
 );
+
+export function convertInputNumberOptionsToNumericInput(
+    inputNumberOptions: ParsedValue<typeof parseInputNumberWidgetV0>["options"],
+): ParsedValue<typeof parseInputNumberWidgetV1>["options"] {
+    return {
+        coefficient: false,
+        rightAlign: inputNumberOptions.rightAlign,
+        size: inputNumberOptions.size,
+        answers: [
+            {
+                status: "correct",
+                value: Number(inputNumberOptions.value),
+                simplify: inputNumberOptions.simplify,
+                message: "",
+                maxError: getMaxError(inputNumberOptions),
+                strict: true,
+                answerForms: getAnswerForms(inputNumberOptions),
+            },
+        ],
+    };
+}
+
+function getMaxError(
+    inputNumberOptions: PerseusInputNumberWidgetOptionsV0,
+): number | undefined {
+    if (!inputNumberOptions.inexact) {
+        return 0;
+    }
+
+    if (inputNumberOptions.maxError == null) {
+        return undefined;
+    }
+
+    return Number(inputNumberOptions.maxError);
+}
+
+const mathFormatsForAnswerType: Record<
+    PerseusInputNumberAnswerType,
+    MathFormat[]
+> = {
+    number: [],
+    decimal: ["decimal"],
+    integer: ["integer"],
+    rational: ["integer", "proper", "improper", "mixed"],
+    improper: ["integer", "proper", "improper"],
+    mixed: ["integer", "proper", "mixed"],
+    percent: ["integer", "decimal", "proper", "improper", "mixed", "percent"],
+    pi: ["pi"],
+};
+
+function getAnswerForms(
+    options: PerseusInputNumberWidgetOptionsV0,
+): MathFormat[] {
+    const value = Number(options.value);
+    const {inexact} = options;
+    const precision = 1e10;
+    const rounded = Math.round(value * precision) / precision;
+
+    const answerType = options.answerType ?? "number";
+    if (answerType === "number" && !inexact && !equalFloats(rounded, value)) {
+        // Disallow decimal answers when the correct answer has more than 10
+        // decimal places. This is for compatibility with legacy input-number
+        // behavior.
+        return ["proper", "improper", "mixed"];
+    }
+
+    return mathFormatsForAnswerType[answerType];
+}
+
+function equalFloats(a: number, b: number): boolean {
+    return Math.abs(a - b) < Math.pow(2, -42);
+}
 
 export const parseInputNumberWidget = versionedWidgetOptions(
     1,
