+++ packages/perseus-score/src/score.test.ts
@@ -1,9 +1,6 @@
 import {
     generateDropdownOptions,
     generateDropdownWidget,
-    generateInputNumberWidget,
-    generateInputNumberOptions,
-    generateInputNumberAnswer,
 } from "@khanacademy/perseus-core";
 import invariant from "tiny-invariant";
 
@@ -11,7 +8,6 @@ import {
     combineScoreWithWidgetScores,
     onlyInvalidScores,
     scorePerseusItem,
-    scorePerseusItemWithInputNumberAsNumericInput,
     scoreWidgetsFunctional,
 } from "./score";
 import {getExpressionWidget, getTestDropdownWidget} from "./util/test-helpers";
@@ -557,228 +553,6 @@ describe("scorePerseusItem", () => {
                 message: null,
             },
         });
-    });
-});
-
-describe("scorePerseusItemWithInputNumberAsNumericInput", () => {
-    it("scores a correctly-answered input-number widget", () => {
-        const renderer = {
-            content: "[[☃ input-number 1]]",
-            widgets: {
-                "input-number 1": generateInputNumberWidget({
-                    options: generateInputNumberOptions({
-                        answers: [generateInputNumberAnswer({value: 42})],
-                    }),
-                }),
-            },
-            images: {},
-        };
-
-        const userInputMap = {
-            "input-number 1": {
-                currentValue: "42",
-            },
-        };
-
-        const score = scorePerseusItemWithInputNumberAsNumericInput(
-            renderer,
-            userInputMap,
-            "en",
-        );
-
-        expect(score).toHaveBeenAnsweredCorrectly();
-        expect(score.widgetScores).toStrictEqual({
-            "input-number 1": {
-                type: "points",
-                total: 1,
-                earned: 1,
-                message: null,
-            },
-        });
-    });
-
-    it("scores an incorrectly-answered input-number widget", () => {
-        const renderer = {
-            content: "[[☃ input-number 1]]",
-            widgets: {
-                "input-number 1": generateInputNumberWidget({
-                    options: generateInputNumberOptions({
-                        answers: [generateInputNumberAnswer({value: 42})],
-                    }),
-                }),
-            },
-            images: {},
-        };
-
-        const userInputMap = {
-            "input-number 1": {
-                currentValue: "99",
-            },
-        };
-
-        const score = scorePerseusItemWithInputNumberAsNumericInput(
-            renderer,
-            userInputMap,
-            "en",
-        );
-
-        expect(score).toHaveBeenAnsweredIncorrectly();
-        expect(score.widgetScores).toStrictEqual({
-            "input-number 1": {
-                type: "points",
-                total: 1,
-                earned: 0,
-                message: null,
-            },
-        });
-    });
-
-    it("scores an empty input-number widget", () => {
-        const renderer = {
-            content: "[[☃ input-number 1]]",
-            widgets: {
-                "input-number 1": generateInputNumberWidget({
-                    options: generateInputNumberOptions({
-                        answers: [generateInputNumberAnswer({value: 42})],
-                    }),
-                }),
-            },
-            images: {},
-        };
-
-        const userInputMap = {
-            "input-number 1": {
-                currentValue: "",
-            },
-        };
-
-        const score = scorePerseusItemWithInputNumberAsNumericInput(
-            renderer,
-            userInputMap,
-            "en",
-        );
-
-        expect(score).toHaveInvalidInput();
-        expect(score.widgetScores).toStrictEqual({
-            "input-number 1": {type: "invalid", message: null},
-        });
-    });
-
-    it("ignores an input-number widget that does not have a placeholder in the content string", () => {
-        const renderer = {
-            content: "",
-            widgets: {
-                "input-number 1": generateInputNumberWidget({
-                    options: generateInputNumberOptions({
-                        answers: [generateInputNumberAnswer({value: 42})],
-                    }),
-                }),
-            },
-            images: {},
-        };
-
-        const userInputMap = {
-            "input-number 1": {
-                currentValue: "",
-            },
-        };
-
-        const score = scorePerseusItemWithInputNumberAsNumericInput(
-            renderer,
-            userInputMap,
-            "en",
-        );
-
-        expect(score).toHaveBeenAnsweredCorrectly({shouldHavePoints: false});
-        expect(score.widgetScores).toStrictEqual({});
-    });
-
-    it("ignores an input-number widget that is not graded", () => {
-        const renderer = {
-            content: "[[☃ input-number 1]]",
-            widgets: {
-                "input-number 1": generateInputNumberWidget({
-                    graded: false,
-                    options: generateInputNumberOptions({
-                        answers: [generateInputNumberAnswer({value: 42})],
-                    }),
-                }),
-            },
-            images: {},
-        };
-
-        const userInputMap = {
-            "input-number 1": {
-                currentValue: "",
-            },
-        };
-
-        const score = scorePerseusItemWithInputNumberAsNumericInput(
-            renderer,
-            userInputMap,
-            "en",
-        );
-
-        expect(score).toHaveBeenAnsweredCorrectly({shouldHavePoints: false});
-    });
-
-    it("scores correctly-answered widgets other than input-number", () => {
-        const renderer = {
-            content: "[[☃ dropdown 1]]",
-            widgets: {
-                "dropdown 1": generateDropdownWidget({
-                    options: generateDropdownOptions({
-                        choices: [{content: "A", correct: true}],
-                    }),
-                }),
-            },
-            images: {},
-        };
-
-        const userInputMap: UserInputMap = {
-            "dropdown 1": {
-                value: 1,
-            },
-        };
-
-        const score = scorePerseusItemWithInputNumberAsNumericInput(
-            renderer,
-            userInputMap,
-            "en",
-        );
-
-        expect(score).toHaveBeenAnsweredCorrectly();
-    });
-
-    it("scores incorrectly-answered widgets other than input-number", () => {
-        const renderer = {
-            content: "[[☃ dropdown 1]]",
-            widgets: {
-                "dropdown 1": generateDropdownWidget({
-                    options: generateDropdownOptions({
-                        choices: [
-                            {content: "A", correct: true},
-                            {content: "B", correct: false},
-                        ],
-                    }),
-                }),
-            },
-            images: {},
-        };
-
-        const userInputMap: UserInputMap = {
-            "dropdown 1": {
-                value: 2,
-            },
-        };
-
-        const score = scorePerseusItemWithInputNumberAsNumericInput(
-            renderer,
-            userInputMap,
-            "en",
-        );
-
-        expect(score).toHaveBeenAnsweredIncorrectly();
     });
 });
